+++ test/support/clients/switch/client.ex
@@ -1,8 +1,0 @@
-defmodule LiveViewNativeTest.Switch.Client do
-  @moduledoc false
-
-  defstruct tags: %{
-    form: "LiveForm",
-    button: "Button"
-  }
-end